move ro-crate metadata processing functions to own module
